+++ Libplanet/Store/BlockSet.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Security.Cryptography;
 using Libplanet.Action;
 using Libplanet.Blocks;
+using Libplanet.Tx;
 
 namespace Libplanet.Store
 {
@@ -64,8 +66,42 @@ namespace Libplanet.Store
 
                 value.Validate(DateTimeOffset.UtcNow);
 
-                // FIXME: A proper mask should be passed.
-                Store.PutBlock(value, default(Address));
+                byte[] prevMask;
+                if (value.PreviousHash is HashDigest<SHA256> prevHash)
+                {
+                    if (Store.GetAddressesMask(prevHash) is Address maskAddr)
+                    {
+                        prevMask = maskAddr.ToByteArray();
+                    }
+                    else
+                    {
+                        // If there is no mask for the previous hash
+                        // (the store made in the older versions of Libplanet
+                        // may lack mask files), make the mask to match to
+                        // all possible addresses.
+                        prevMask = Enumerable.Repeat((byte)0xff, Address.Size)
+                            .ToArray();
+                    }
+                }
+                else
+                {
+                    prevMask = new byte[Address.Size];
+                }
+
+                // Create a next mask.
+                var mask = new BitArray(prevMask);
+                foreach (Transaction<T> tx in value.Transactions)
+                {
+                    foreach (Address address in tx.UpdatedAddresses)
+                    {
+                        // BitArray.Or() manipulates itself.
+                        mask.Or(new BitArray(address.ToByteArray()));
+                    }
+                }
+
+                var maskBuffer = new byte[Address.Size];
+                mask.CopyTo(maskBuffer, 0);
+                Store.PutBlock(value, new Address(maskBuffer));
             }
         }