+++ CHANGES.md
@@ -17,6 +17,9 @@ To be released.
     mined by `BlockChain<T>.MineBlock()`.
  -  Fixed a bug that `TurnClientException` had been thrown by Swarm when a STUN
     nonce is stale.  [[#193]]
+ -  Fixed `BlockChain<T>.GetStates()` had descended to the bottom
+    (i.e., the genesis block) where a given `Address` referes to
+    a nonexistent account (i.e., never used before).  [[#189]]
  -  Added `GetAddressesMask(HashDigest<SHA256>)` method to `IStore` interface
     and its all implementations.  [[#189]]
  -  The signature of `IStore.PutBlock<T>(Block<T>)` method was changed to